# api/openapi.yaml
openapi: 3.0.3
info:
  title: remoses
  version: 0.1.0
  description: |
    Remote control of amateur radio transceivers.

    remoses runs next to the radios and exposes frequency, mode, filter, power,
    PTT and CW sending over HTTP, with a WebSocket stream of state changes.

    ## Authentication

    HTTP Basic on every request, over TLS. Authorisation is per instance: any
    authenticated user may use any radio.

    ## Locking

    Exclusive control is per radio. Acquire a lock, then pass its token in the
    `X-Remoses-Lock` header on every state-changing request. The lease slides:
    any accepted command renews it. Reads and the WebSocket need no lock.

    If a lock expires or is released while the radio is transmitting, remoses
    drops PTT and flushes the CW queue. A client that crashes mid-transmission
    must not leave a carrier up.
servers:
  - url: /api/v1
security:
  - basicAuth: []
...
paths:
...
  /radios/{radioId}/cw:
    parameters:
      - $ref: "#/components/parameters/RadioId"
    get:
      tags: [cw]
      operationId: getCW
      summary: CW queue status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CWStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

    post:
      tags: [cw]
      operationId: sendCW
      summary: Enqueue CW text
      description: |
        Text is in remoses canonical form, which is rig-neutral: a prosign is
        `^` followed by the letters to be keyed with no inter-character spacing,
        such as `^AR` or `^SK`. remoses translates that into whatever the target
        rig speaks — Icom's `^` marker, Kenwood's single-symbol substitutions, or
        a local element table.

        Returns as soon as the text is queued; it does not wait for it to be
        sent. Characters the radio cannot key are rejected with 422 rather than
        silently mangled by the rig.
      parameters:
        - $ref: "#/components/parameters/LockToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CWRequest" }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CWAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/LockConflict" }
        "422":
          description: |
            Unsendable character, or CW not supported by this radio. The problem
            document names the offending character, its offset, and the accepted
            charset.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "423": { $ref: "#/components/responses/Locked" }
        "503": { $ref: "#/components/responses/Disconnected" }

    delete:
      tags: [cw]
      operationId: abortCW
      summary: Abort sending
      description: |
        Drops the local queue and tells the rig to stop. Both halves matter: up
        to a full buffer may already be inside the radio.
      parameters:
        - $ref: "#/components/parameters/LockToken"
      responses:
        "204": { description: Aborted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/LockConflict" }
        "423": { $ref: "#/components/responses/Locked" }
  /radios/{radioId}/lock:
    parameters:
      - $ref: "#/components/parameters/RadioId"
    get:
      tags: [lock]
      operationId: getLock
      summary: Who holds the lock
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LockState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

    post:
      tags: [lock]
      operationId: acquireLock
      summary: Acquire exclusive control
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LockRequest" }
      responses:
        "201":
          description: Acquired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lock" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Steal requested but disabled by configuration
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/LockConflict" }

    delete:
      tags: [lock]
      operationId: releaseLock
      summary: Release
      description: |
        Releasing while transmitting drops PTT and flushes the CW queue.
      parameters:
        - $ref: "#/components/parameters/LockToken"
      responses:
        "204": { description: Released }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "423": { $ref: "#/components/responses/Locked" }
  /radios/{radioId}/lock/renew:
    parameters:
      - $ref: "#/components/parameters/RadioId"
    post:
      tags: [lock]
      operationId: renewLock
      summary: Heartbeat the lease
      description: |
        Any accepted command renews the lease already. This exists for a client
        holding the radio while its operator thinks, without issuing commands.
      parameters:
        - $ref: "#/components/parameters/LockToken"
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lock" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/LockConflict" }
        "423": { $ref: "#/components/responses/Locked" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    RadioId:
      name: radioId
      in: path
      required: true
      description: Stable radio identifier from the configuration file.
      schema: { type: string, example: ic7610 }
    LockToken:
      name: X-Remoses-Lock
      in: header
      required: false
      description: |
        Lock token from POST /radios/{radioId}/lock. Required for every
        state-changing request when locking is enabled. A `remoses_lock_{id}`
        cookie is accepted as a fallback for browsers; the header wins.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such radio
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    LockConflict:
      description: The lock is held by another user
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Locked:
      description: No lock held, or the token is expired or invalid
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Disconnected:
      description: The radio is not currently connected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      description: RFC 9457 problem detail, with a radio_id extension member.
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        radio_id: { type: string }
        locked_by: { type: string }
        expires_at: { type: string, format: date-time }
        character: { type: string }
        offset: { type: integer }
        charset: { type: string }
...
    CWStatus:
      type: object
      required: [busy, queued, wpm]
      properties:
        busy: { type: boolean }
        queued: { type: integer, description: Characters not yet keyed }
        wpm: { type: integer }
        est_remaining_ms: { type: integer }
...
    LockRequest:
      type: object
      properties:
        force:
          type: boolean
          description: Steal a held lock. Permitted only when enabled in configuration.
    Lock:
      type: object
      required: [token, expires_at, ttl_seconds]
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        ttl_seconds: { type: integer }
    LockState:
      type: object
      required: [held]
      properties:
        held: { type: boolean }
        holder: { type: string }
        expires_at: { type: string, format: date-time }
        is_mine: { type: boolean }
    CWRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Canonical form; `^AR` style prosigns.
          example: CQ TEST DE OH2XYZ ^AR
        wpm: { type: integer }
        mode:
          type: string
          enum: [append, replace]
          default: append
    CWAccepted:
      type: object
      required: [queued_chars]
      properties:
        queued_chars: { type: integer }
        position: { type: integer }
        est_duration_ms: { type: integer }
...